add support for extended boot delays if signalled by the application
Added a mechanism for the application to signal to the bootloader
that it should wait an additional 30 seconds before booting into
the application. This is accomplished by setting a special magic
value and command value at a predefined location in SRAM. Since
SRAM contents are preserved across soft resets, the bootloader
checks for the presence of this value and extends the delay
if found. This feature is intended for the case where the
Ethernet link takes longer than usual to come up for some
reason.

Additionally, a random 0-255 millisecond delay was added
in the bootloder init sequence to prevent a broadcast storm
in the case where there's a large number of these devices on
the network all being powered on at the same time.